requirements.txt
torch>=2.0.0
streamlit>=1.25.0
Pillow>=10.0.0
numpy>=1.24.0